requirements.txt
streamlit
opencv-python
pyzbar
pillow
numpy
